added compilationoptions
