+++ src/Buckle/Buckle.Tests/Assertions.cs
@@ -23,7 +23,10 @@ internal static class Assertions {
     /// <param name="expectedValue">Expected result.</param>
     internal static void AssertValue(string text, object expectedValue) {
         var syntaxTree = SyntaxTree.Parse(text);
-        var compilation = Compilation.CreateScript(null, syntaxTree);
+        var compilation = Compilation.CreateScript(
+            new CompilationOptions(BuildMode.Independent, true, false), null, syntaxTree
+        );
+
         var variables = new Dictionary<VariableSymbol, EvaluatorObject>();
         var _ = false;
         var result = compilation.Evaluate(variables, ref _);
@@ -43,7 +46,10 @@ internal static class Assertions {
     /// <param name="exceptions">Expected exception(s) thrown.</param>
     internal static void AssertExceptions(string text, ITestOutputHelper writer, params Exception[] exceptions) {
         var syntaxTree = SyntaxTree.Parse(text);
-        var compilation = Compilation.CreateScript(null, syntaxTree);
+        var compilation = Compilation.CreateScript(
+            new CompilationOptions(BuildMode.Independent, true, false), null, syntaxTree
+        );
+
         var _ = false;
         var result = compilation.Evaluate(new Dictionary<VariableSymbol, EvaluatorObject>(), ref _);
 
@@ -77,7 +83,10 @@ internal static class Assertions {
         if (syntaxTree.diagnostics.Errors().Any()) {
             tempDiagnostics.Move(syntaxTree.diagnostics);
         } else {
-            var compilation = Compilation.CreateScript(null, syntaxTree);
+            var compilation = Compilation.CreateScript(
+                new CompilationOptions(BuildMode.Independent, true, false), null, syntaxTree
+            );
+
             var _ = false;
             var result = compilation.Evaluate(new Dictionary<VariableSymbol, EvaluatorObject>(), ref _);
             tempDiagnostics = result.diagnostics;
@@ -127,7 +136,7 @@ internal static class Assertions {
     /// <param name="buildMode">Which emitter to use.</param>
     internal static void AssertText(string text, string expectedText, BuildMode buildMode) {
         var syntaxTree = SyntaxTree.Parse(text);
-        var compilation = Compilation.Create(buildMode == BuildMode.CSharpTranspile, syntaxTree);
+        var compilation = Compilation.Create(new CompilationOptions(buildMode, false, false), syntaxTree);
         var result = compilation.EmitToString(buildMode, "EmitterTests");
 
         Assert.Empty(compilation.diagnostics.Errors().ToArray());
