+++ src/Buckle/Buckle/CodeAnalysis/Binding/Semantics/Operators/BoundTernaryOperatorKind.cs
